Refector: pagination component refectoring
